Integrating the homework3 into lab0-c
Add the contents related to MCTS in the ttt project to the lab0-c,
and modify the Makefile to compile the newly added files.
Also, update the pre-hook content to resolve the warnings issued
by cppcheck so that commits can pass.
Add hlist for zobrist.
Handling the issue of using const variables for game and mt19937-64.